requirements.txt
Flask==3.1.0
jira==3.8.0
python-dotenv==1.1.0
Werkzeug==3.1.3
